+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add option in `nessai.reparameterisations.Angle` to set `scale=None`, the scale is then set as `2 * pi / angle_prior_range`.
 - Add `'periodic'` reparameterisation that uses `scale=None` in `nessai.reparameterisations.Angle`.
 - Add the `use_default_reparameterisations` option to `FlowProposal` to allow the use of the default reparameterisations in `GWFlowProposal` without specifying any reparameterisations.
+- Add `chi_1`, `chi_2` and `time_jitter` to known parameters in `GWFlowProposal` with corresponding defaults.
 
 ### Changed
 
@@ -36,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Model.names` and `Model.bounds` are now properties by default and their setters include checks to verify the values provided are valid and raise errors if not.
 - Logger now has propagation enabled by default.
 - `FlowProposal.configure_reparameterisations` can now handle an input of `None`. In this case only the default reparameterisations will be added.
+- Changed default reparameterisation for gravitational-wave parameters `a_1` and `a_2` to `'default'`.
 
 ### Fixed